requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.26.2
torch==2.1.1
pgvector==0.2.4
